backend/requirements.txt
asgiref==3.8.1
Django==5.1.4
django-filter==2.4.0
django-cors-headers==4.6.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
pytz==2021.1
sendgrid==6.11.0
sqlparse==0.4.1
mysqlclient==2.2.6
